refactor(consensus): move files under types folder in accordance with original structure
