+++ src/noir/consensus/block_executor.h
@@ -6,10 +6,10 @@
 #pragma once
 #include <noir/consensus/abci_types.h>
 #include <noir/consensus/app_connection.h>
-#include <noir/consensus/event_bus.h>
 #include <noir/consensus/store/block_store.h>
 #include <noir/consensus/store/state_store.h>
 #include <noir/consensus/types.h>
+#include <noir/consensus/types/event_bus.h>
 #include <noir/consensus/types/events.h>
 #include <noir/tx_pool/tx_pool.h>
