+++ src/noir/consensus/privval/file.h
@@ -8,7 +8,7 @@
 #include <noir/common/refl.h>
 #include <noir/common/types.h>
 #include <noir/consensus/crypto.h>
-#include <noir/consensus/priv_validator.h>
+#include <noir/consensus/types/priv_validator.h>
 #include <noir/p2p/protocol.h>
 #include <fc/io/json.hpp>
 #include <filesystem>
