+++ src/noir/consensus/types/vote.h
@@ -4,8 +4,8 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 //
 #pragma once
-#include <noir/consensus/block.h>
-#include <noir/consensus/validator.h>
+#include <noir/consensus/types/block.h>
+#include <noir/consensus/types/validator.h>
 #include <noir/p2p/protocol.h>
 #include <noir/p2p/types.h>
